changes required by moving URI to a separate package
