+++ app/src/main/java/com/waz/zclient/AppEntryActivity.java
@@ -33,7 +33,7 @@ import com.waz.api.ImageAsset;
 import com.waz.api.ImageAssetFactory;
 import com.waz.api.LoadHandle;
 import com.waz.api.Self;
-import com.waz.utils.AndroidURIUtil;
+import com.waz.utils.wrappers.AndroidURIUtil;
 import com.waz.zclient.controllers.navigation.NavigationControllerObserver;
 import com.waz.zclient.controllers.navigation.Page;
 import com.waz.zclient.controllers.tracking.screens.ApplicationScreen;
