+++ app/src/main/java/com/waz/zclient/controllers/notifications/ShareSavedImageActivity.java
@@ -20,8 +20,8 @@ package com.waz.zclient.controllers.notifications;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
-import com.waz.utils.AndroidURI;
-import com.waz.utils.URI;
+import com.waz.utils.wrappers.AndroidURI;
+import com.waz.utils.wrappers.URI;
 import com.waz.zclient.BaseActivity;
 import com.waz.zclient.notifications.controllers.ImageNotificationsController;
 import com.waz.zclient.utils.IntentUtils;
