+++ app/src/main/java/com/waz/zclient/controllers/singleimage/ISingleImageController.java
@@ -20,7 +20,7 @@ package com.waz.zclient.controllers.singleimage;
 import android.view.View;
 import com.waz.api.Message;
 import com.waz.api.User;
-import com.waz.utils.URI;
+import com.waz.utils.wrappers.URI;
 
 public interface ISingleImageController {
     void addSingleImageObserver(SingleImageObserver observer);
