+++ app/src/main/java/com/waz/zclient/controllers/singleimage/SingleImageObserver.java
@@ -19,7 +19,7 @@ package com.waz.zclient.controllers.singleimage;
 
 import com.waz.api.Message;
 import com.waz.api.User;
-import com.waz.utils.URI;
+import com.waz.utils.wrappers.URI;
 
 public interface SingleImageObserver {
     void onShowSingleImage(Message message);
