+++ app/src/main/java/com/waz/zclient/pages/main/sharing/ImageSharingPreviewFragment.java
@@ -25,7 +25,7 @@ import android.widget.FrameLayout;
 import com.waz.api.IConversation;
 import com.waz.api.ImageAsset;
 import com.waz.api.ImageAssetFactory;
-import com.waz.utils.URI;
+import com.waz.utils.wrappers.URI;
 import com.waz.zclient.BaseActivity;
 import com.waz.zclient.R;
 import com.waz.zclient.controllers.drawing.DrawingController;
