+++ build.gradle
@@ -73,7 +73,8 @@ ext {
     playServicesVersion = '10.0.1'
 
     audioVersion = System.getenv("AUDIO_VERSION") ?: '1.209.0@aar'
-    zMessagingDevVersionBase = '97.1439'
+    //zMessagingDevVersionBase = '97.1439'
+    zMessagingDevVersionBase = '97-SNAPSHOT'
     zMessagingDevVersion = "${zMessagingDevVersionBase}@aar"
     // Release version number must be like this X.0(.Y)
     zMessagingReleaseVersion = System.getenv("ZMESSAGING_VERSION") ?: '97.1.339@aar'
